refactor: refactorimg Usage Insights in www
