+++ www/src/pages/insights/utils/getEmptyMessage.tsx
@@ -1,0 +1,3 @@
+export default function getEmptyMessage(text: string) {
+  return `Currently there are no ${text} usage yet`;
+}
