+++ www/src/pages/insights/utils/usagePropTypes.tsx
@@ -1,0 +1,7 @@
+import PropTypes from 'prop-types';
+
+const usagePropTypes = {
+  data: PropTypes.arrayOf(PropTypes.string).isRequired,
+};
+
+export default usagePropTypes;
