+++ www/src/types/types.ts
@@ -1,0 +1,46 @@
+export interface IInsightsContext {
+  paragonTypes: {
+    [key: string]: string
+  },
+  isParagonIcon: Function,
+}
+
+export interface IUsage {
+  filePath: string,
+  line: number,
+  column: number,
+  version: string,
+}
+
+export interface IComponentUsageData {
+  componentUsageCount: number,
+  folderName: string,
+  name: string,
+  repositoryUrl: string,
+  usages: IUsage[],
+  version: string,
+}
+
+export interface IDependentProjectsUsages extends Omit<IDependentUsage, 'count'> {
+  version: string,
+  name: string,
+  repository: { type: string, url: string },
+  folderName: string,
+}
+
+export interface IDependentUsage {
+  version?: string,
+  name?: string,
+  repository?: { type: string, url: string },
+  repositoryUrl?: string,
+  count: number,
+  folderName?: string,
+  usages: {
+    [key: string]: IUsage[],
+  },
+}
+
+export interface IComponentUsage {
+  name: string,
+  componentUsageInProjects: IComponentUsageData[],
+}
